+++ app/(tabs)/e-bilet.tsx
@@ -145,7 +145,6 @@ const styles = StyleSheet.create({
     flex: 1,
     backgroundColor: "#6C78E6",
     alignItems: "center",
-    paddingBottom: 70
   },
   scrollContent: {
     alignItems: "center",
