+++ app/_layout.tsx
@@ -35,9 +35,6 @@ export default function RootLayout() {
         <Stack>
           <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
           <Stack.Screen name="+not-found" />
-          <Stack.Screen name="portal-dokumentow" options={{ headerShown: false,  }} />
-          <Stack.Screen name="materialy-marketingowe" options={{ headerShown: false,  }} />
-          <Stack.Screen name="informacje-targowe" options={{ headerShown: false,  }} />
         </Stack>
         <StatusBar style="auto" />
       </ThemeProvider>
